Add files via upload
I've been working on this project for a while now. I want it to be the best of its kind eva! Its an app called *Aeon Media Player*. I even want it to be much like VLC! I do not take any credit for the photos/images/icons used in this app, which were downloaded from the internet. In the final version, all the credits will be listed out!